move resize-observer and intersection service to utilities
