+++ packages/web-components/fast-foundation/src/utilities/intersection-service.ts
